fix indexing and border

+++ src/utils/plot_helpers.py
@@ -39,7 +39,7 @@ def put_palette_1d(image_arr, palette, ids=None):
     else:
         print(len(palette), len(ids))
         print(palette, ids)
-    new_image_arr = np.zeros(image_arr.shape, dtype=np.uint8)
+    new_image_arr = np.zeros(image_arr.shape, dtype=np.int16)
     for i, idx in enumerate(ids):
         new_image_arr[image_arr == idx] = palette[i]
     return new_image_arr
